clean up repository and import into Git
This is just a port of the SVN repository contents, I haven't relocated
any files yet. Everything inside the jes/ folder, as well as the
Release Build Tools folder and a couple minor documentation files, was
copied over, and I've tested that the Makefile works (at least, `make`,
`make tests`, and `make run`).

Most of this will probably be moved around to be better sorted, and
the various methods of building and launching will be consolidated.
In the meantime, a new `README.md` describes where everything lives.