correct "Progam" to "Program"
The "Loading Progam" text has been a long-standing and embarrassing
bug, and now it has been corrected. (There was also an instance of
"Progam" in one of the auxiliary help files. I can't find out how you
get to that help file to begin with, but it's still worth fixing.)