reorganize the source into jes, dependencies, and tests
Pretty much everything has been moved into the jes, dependencies, and
tests directories, depending on its purpose. The layout is fully
described in the README, but the short of it is that all of JES's
original code/content lives in jes, everyone else's code lives in
dependencies, and the unit tests live in tests (and don't get
distributed).

The Makefile can launch JES in the new setup, but none of the other
launchers can yet. (Fortunately, I have a pretty good idea of how to
port them all, and most of my thoughts are written down in the
working-on-jes directory.)

The new launching structure introduces the concept of the jes.home
directory, which is managed by the JESResources Java class. This is the
directory that the JES code looks for everything in at runtime: images,
Python modules, help, whatever. The nice thing about this is that now,
you don't have to have a specific current directory for JES to work.

Also, some random unused JAR's, batch files, and stuff were
intentionally left behind.