prevent two >>>'s from being printed at boot
A bug suddenly manifested during the Jython upgrade that caused >>> to
appear in the command line (after the actual >>> prompt) upon JES
startup. It was probably caused by some of the Command Window changes
I had to make during the Jython upgrade.

Further investigation revealed that a similar bug has appeared in the
past, and it was fixed by removing a call to resetConsole() in JESUI,
leaving a call to setText('>>>') in JESCommandWindow. I reversed the
order of the removals, so that resetConsole() is called without a
setText() first. I'm not sure why this works...I have hypotheses,
but nothing definite.

Regardless of the cause, the bug has been repaired as far as I can
tell: only one >>> is printed, and it can't be backspaced over, and
the command window works in other areas.