print information from exceptions, not classes
The ExceptionRecord class was, for some inscrutable reason, passing
exc_type to the method that finds exception descriptions instead of
exc_value, so all the descriptions with placeholders broke with the
upgrade to 2.5. (I don't know how it worked before, which seems to
be a theme with my changes to JES.)

Also some minor cleanups that I just happened to do while working on
this problem.