fix rounding in SimpleTurtle distance calculation
Due to float imprecision, occasionally a pixel will be dropped from the
distance to move - (int) uses truncation, but to get the most accurate
estimate, a round-by-half method is necessary. So, calls to Math.round
were added.

Thanks to Christine Alvarado from UC San Diego for discovering this bug
and finding a fix. (It's been too long since Calculus II for me to
develop a proper test case, so I'm relying on her confirmation of the
fix.)