port build system from Make to Ant
This allows it to work on Windows as well as Linux, without pain.
(The first thing I tried was porting the Makefile to a .bat file,
and that got complicated fast.)

This also activated the shell script. The batch file doesn't work yet,
because I don't have a development environment on Windows fully
configured, but that should happen soon.

Distribution releases will also be made through Ant soon.