make tests executable from within JES by adding __file__
It's not the most user-friendly interface for running the tests, but it
is nice to know that they _can_ run inside JES. It can even run with
the watcher enabled, but it slows down by an absolutely ridiculous
amount.

The only change _required_ was to add the __file__ variable to the
execution context, so the tests could find their resources. However,
the structure of the test file was adjusted to make it easier to run
from within JES and see the results. (Also some print statements were
removed.)