make working jes.bat for Windows
jes.bat is just a port of jes.sh to a batch file, which is much easier
than porting the Makefile to a batch file. It configures all the same
system properties and environment variables, and uses JAVA_HOME the same
way. (Though soon it will need to be modified to deal with vendored
JRE's.)