make a .tar.gz package for Linux (and Mac)
'ant release-linux' will now build a .tar.gz package that you can
unpack and run directly. (It should work on any Unix.)

The release process is controlled in releases/build-releases.xml,
which also provides 'ant release' to make all the releases at once
and 'ant clean-release' to wipe them out. Continued testing is
necessary to determine which releases can be built on which platforms.

This gets rid of the 'packaging' directory and moves everything that
you don't need when you're just developing on JES into 'releases',
with the previous contents of 'packaging' living in
'releases/resources'.