add a Windows zip-nojava build
This just puts all the needed files in a zip file, including the .exe
and the .bat. As long as you have a JRE configured, this is nice and
portable. It's split into a -contents task and an actual release task,
because we'll reuse it for the other three kinds of Windows build.