add a Windows installer, using NSIS
The installer is super-basic right now, but it installs and uninstalls
cleanly, which is all we really need. It's only tested on Wine - once I
get local admin access on my Windows box, I'll test it out there too.

The installer is built by Ant, but you need to define the "makensis"
parameter with the path to makensis. (Instructions to that effect are
included in README.md.)