fix .bat file's classpath handling
I forgot to enclose %classpath% in quotes (too much fish programming, I
guess). It worked in the git checkout because
C:\Users\mfrazier\Documents\GitHub\jes has no spaces in it, but
C:\Program Files\JES does.

To facilitate debugging this, I switched the "javaw" call to "java," and
I removed the "start" command. JES.exe already takes care of launching
the .bat in the background, so it's OK if the .bat file itself launches
Java in the foreground and keeps a prompt window open.