make some modifications to the installer
The installer that we already had didn't work perfectly. Besides
containing the bad .bat file, it wasn't uninstalling everything, and the
uninstaller registry key wasn't working. I made a few changes, and I'm
not sure exactly which of them were necessary, but it works now and
that's the important thing.