make the installer install a JRE automatically
In order to improve the user experience, the .exe installer now includes
a ZIP file of the JRE. An NSIS section controls whether the JRE gets
unpacked. When it runs, it checks whether a JRE is installed, and if
not, hard-codes the JRE option to "on."

This makes the installer four times as large, but should improve the UX
greatly. (And if Java is already installed, it won't actually unpack the
JRE, so the extra space will only appear in the Downloads folder.)