rename JESConfig.txt to .properties in launchers
While I'm moving the location of the config file, I decided it would be
a good idea to change the format to a .properties file to make the file
less fragile and easier for users to understand.

This commit implements the changes in the launchers, which are really
just "change JESConfig.txt to JESConfig.properties." The code itself
still stores it in JESConfig.txt in the home directory, but the next
commit will implement the new configuration handling.