store settings in .properties file
This new configuration format is much more future-proof and resistant
to accidental changes. The new JESConfig class is interface-compatible
with the old, so there shouldn't be any incompatibilities resulting
from this change.

JESConfig includes enough information about the old format to read it
and transform it into the new format, after which the new .properties
file will be used automatically.

Also, some code was added to JESProgram to report load errors and
migrated settings.