add release index and testing kit for testers
To facilitiate testing and downloading JES development builds,
build-releases.xml creates an index.html file in each release directory
with a list of all available releases (regardless of whether they were
built!) for users to browse. This means you can just rsync the entire
directory to a server somewhere and give people the URL.

It also zips up the tests, demo scripts, and (in the future) other
stuff to create a testing kit ZIP file. This allows users to run the
JES tests on their own platforms.