add a preliminary UI test plan
This plan exercises *most* of the core of JES (and especially the
Watcher and speed slider), but it doesn't test most of the ancillary
functionality (like the Options dialog), or test the interactions
between different media components/JES functions. Additional test
components will probably be added later. In the meantime, this is
sufficient to reproduce the problems we have been experiencing so far.