force Unicode strings to native strings in media.py
Presumably as a consequence of the Jython 2.5 upgrade, the string
return values from pickAFile(), requestString(), and friends were being
represented as Unicode, and therefore prefixed with 'u'. Mark realized
that this was a BC break for our screenshots, and to avoid having to
explain Unicode in the text so early, we instead force all strings to
native strings. I don't know how native strings actually work in Jython
or the Java I/O system, but so far it seems to work well enough.