provide flag to ensure Python thread shutdown
On Mark's laptop (Mac OS X/JRE 6), if a Thread.stop() signal arrived
while the thread was asleep, the thread wouldn't actually stop. The
sleep would interrupt, but the thread would continue running.

So, JESThread now includes a flag that is set to True when the thread
is stopped. Before and after JESDebugger sleeps, it checks the flag,
and raises ThreadDeath if one was not already raised by the .stop().