override exception message for ThreadDeath
java.lang.ThreadDeath is kind of a scary message to see when you click
the "Stop" button. So, now, instead of just having a message that gets
buried below the death-having traceback, JESExceptionRecord completely
overrides the dislay logic for ThreadDeath exceptions, to print out a
simple message saying that the program stopped.