remove annoying exception wrapper in JESUI
The exception handler in the JESUI constructor is one of the worst
ideas I've seen in all of JES, and I've seen plenty of bad ideas in
JES. If any exception happens during that constructor, JES can't
start. However, the exception handler eats the exception and causes
the splash window to hang open forever, instead of allowing the
program to terminate.