rewrite CommandDocument{Filter,Listener} in Java
I'm currently trying to debug the keyboard lockup issue, and I suspect
it has to do with event listeners taking too long. To help offset this,
all the code that gets invoked on every command window change has been
rewritten into Java: the Filter and Listener are both Java, and the
History inherits from a Java class that manages the current buffer so
that it doesn't have to call into Python.

(I _think_ it's helped, but I haven't been quite scientific about how
I tested this.)