rewrite most of the interpreter/debugger
Another round of jes/python/JESWhatever.py files have been deleted,
as the interpreter and debugger were rewritten to live in jes.core.
The new code provides a clean separation between the interpreter core
and the GUI code that manipulates it and displays information from it.
Besides cleaning up the code in general, it also makes it easier for
JES to be threadsafe, as the GUI-touching code can be wrapped in
@threadsafe.

To support such a separation, the Blinker library is vendored. Blinker
is a nice library for sending signals, which most of the GUI code uses
to bind itself to events on the Python core code.

The debugger/watcher still works, but the control panel for the
debugger doesn't exist yet, so you can't watch any variables and the
speed is hard-coded to 2 lines per second.

Also, I think these changes have successfully removed the bug where JES
occasionally refuses to start under Linux OpenJDK 1.7.0_60.

This marks the beginning of "Operation This Is Not Java," whose end
goal is the destruction of all the JESWhatever.py files as features rae
moved into jes/python/jes and properly isolated.