create different themes for the Command Window
Right now, there are two themes: "Modern" and "JES 4.3".
You can call setTheme() on the command window to switch the themes,
and AFAICT theme switching is totally seamless. There's no UI for it
yet, though.

(JES 4.3 will be the default set in the JESConfig system, and it's
designed to look almost 100% like JES 4.3. The only difference is
that it uses Monospaced and not the system default font for prompts.
However, "Modern" is set as the internal default.)