move editor focus tracking into JESUI
Since it's JESUI that cares about whether the editor pane or command
window is focused, the focus-tracking code has been moved there.
So, neither JESEditor nor CommandWindowPane are FocusListeners anymore.

This also restores the menu functionality (copy, cut, paste, insert
function name) that was lost when the command window was rewritten.

(Also, instead of FocusOwner, it's been changed to focusedEditor.)