cap number of records in watcher
In order to prevent Python from allocating a potentially infinite
number of records, JES 4.3 implemented a cap on the number of records.
I hadn't reimplemented this yet. (Fortunately I hadn't written any
infinite loops yet, because that would have undoubtedly killed my
laptop.)

Now, a collections.deque is used to store the records, and once the
record limit is hit, records are popped from the left before appending
the new one. The number of records are reported to the table listener,
which uses it to add a message at the beginning, indicating that
records were removed.