add theme-changing UI for the command window
This takes the form of another option in the Options window.
(Yes, this pollutes the JESUI.py even more. I'm willing to accept that
in the short term for getting this feature implemented.)

Along with it, some new themes were added. There are simple White on
Black and Black on White themes, as well as a pair of prototype themes
optimized for colorblind users. (The colorblind themes are not final
yet, as I'm having difficulty determining appropriate colors.)