change skin now changes watcher and help skins
Before, the watcher and help screens would almost always display with
the "Metal" skin, regardless of the current skin setting. This is
because they aren't visible when the skin was set during startup, and
after startup, they would only have their skin changed when they were
open and the user visited the Options dialog.

This has revealed that the Watcher's layout code doesn't quite work
properly on all looks and feels. I'm going to research ways to make
them fit better on the more native controls.