adjust debugger panel appearance on non-Metal L&F
Instead of fixing a particular size for the buttons, the border size
is set to (0, 0, 0, 0). The buttons stay square because the slider's
pushing against them.

(This means they have borders now, which Mark _may_ consider a UI
break, but I think that's within the normal set of UI differences
you'd see with different skins and platforms.)

It also removes the speed value floating over the slider in GTK+ mode.
The number doesn't have meaningful units anyway...