print a \n as part of sys.displayhook
The CommandWindow displayhook implementation doesn't print a newline
after the value. This isn't normally a problem because each statement
usually only displays one result, and the prompt() method inserts a
newline when the prompt is displayed, but if a simple expression is
used as a statement in a for loop, all the results would be printed
directly after one another.

So, now displayhook adds a \n after the repr() of the value.