move JES Python startup into separate file
Now, the JES startup code lives in jes/python/jes/__main__.py, which
JESstartup invokes using the -m flag to Jython. This gives us more
flexibility in the early launch stages.

(This code was originally written to experiment with calling the whole
JESProgram constructor on the EDT. That didn't work out, but I still
think it's a good change.)