add a callAndTime function for benchmarking
I originally added this function just for my own personal use, but
students may find it helpful as well, since it gives you really
accurate timing. It just prints out the amount of time it takes to call
a function.

There's also a pixel-benchmarks.py script that uses the new JES --run
mode and the new callAndTime function to test the speed of different
pixel operations. I plan to do some refactoring of pixel operations
soon, so this will come in handy.