link command window caret visibility to focus
My _last_ attempt to fix the caret visibility issues turned out to be
an epic failure, as starting JES resulted in both carets being useful.
(I should not have committed it.) This time, I added a FocusListener
that sets the caret.visible property appropriately whenever focus is
gained or lost from the command window. Testing reveals that this works
and was probably what I should have done in the first place.