run tests through JES launcher instead of Ant
Before this commit, information about how to launch JES was in each
of the three launchers, plus the build.xml file -- and it took up a
substantial part of build.xml. Now, however, JESstartup has been
configured with a passthrough that allows you to directly access the
Jython shell hidden within.

So, the build.xml file now uses that --jython switch to run the tests,
instead of duplicating all the logic for launching Java.