clean up launchers and add JESEnvironment.{sh,bat} file
To make configuring the JES environmental settings easier, the
launchers now look for a JESEnvironment.sh (or .bat, if on Windows)
file that lives in the same directory as JESConfig.properties.
It will then source before picking a Java and whatnot.

In order to keep the diffs between the three launchers minimal,
I rearranged them a bit so that everything happens in the same order.