clean up command window themes
In the area of naming, all of them use "on White" or "on Black" now,
and "JES 4.3" has been renamed to "Old JES (Yellow on Black)".

In addition, a new "Standard on White" theme was added. It uses a color
mix similar to Standard, and was constructed in a similar manner to
Standard (I pick colors and see which ones I think work well together).