move SessionWrapAround responsibility to Pixel
Now, instead of JESConfig having to track session-specific changes to
the pixel wraparound behavior, the setting is stored on Pixel, and
examined/altered with Pixel static methods.

There are a couple of reasons for this change. First, it's weird that
JESConfig has code to deal _specifically_ with changing one setting.

Second, it established an unnecessary dependency between Pixel and
JESConfig, as well as the tests and JESConfig.

Third, because of how frequently this setting is consulted in hot
loops, having it as an static variable in the class that consults it
should speed performance. (Preliminary benchmarks suggest that this
is the case.)