remove a bunch of old Python files
Debug, bar, and swing appear to be old leftovers that were used
internally at one point and are now obsolete or otherwise useless.

inspect and dis are copies of modules from the Jython standard
library. They don't appear to be used, and even if they were,
we have functioning copies in the stdlib now.

tabnanny was used as the basis for JESTabnanny. The only differences
between tabnanny and JESTabnanny is that JESTabnanny's CLI code has
been removed, and the only differences between tabnanny here and in
the stdlib is that ours has better formatting.

JESDebugWindowWrapper is simply not used anymore.