use extended set syntax to fix Win64 bugs
On 64-bit Windows, JES gets installed to "Program Files (x86)".
The batch interpreter can mess up when you use a variable containing
parentheses in a parenthesized block.

I was testing on a 32-bit machine, so it didn't have the parens,
and I didn't notice this before. Now that I have a 64-bit machine to
test on, I have fixed this.