fix OS version detection on Windows
On Windows, the consumer-known version is included in the os.name
property, with Microsoft's internal version number in os.version.
The internal version number isn't very useful, and is also kind of
weird (for example, Windows 7's os.version is 6.1). So, it now checks
for "Windows " in os.name, and doesn't print os.version in that
case.