fix encoding issues in build on Windows
There were a couple of situations where Windows was interpreting files
as Windows-1252 instead of UTF-8. Most notably, NSIS assumes everything
is in 1252, and so the copyright file was displaying assorted other
symbols instead of the Copyright symbol. (This was fixed by transcoding
the Copyright and Readme files from UTF-8 to Windows-1252 during build.)

But the JVM's default encoding for source files and data files on
Windows is also 1252, which caused similar symbol misinterpretations
when JES was compiled or run on Windows. An encoding="UTF-8" directive
was added to the Ant <compile> task, and all launchers had
-Dfile.encoding=UTF-8 added to ensure that JES is using UTF-8 whenever
it needs to load its own stuff.