associate JES with .py files on Windows
When JES is installed, it enrolls itself in the registry as a valid
opener of .py files. If the system doesn't know anything about .py files
yet, it adds some basic information to the registry and sets itself as
the default. (This is part of the installer, not JES itself.)

What happens in the presence of other .py editors hasn't been fully
tested (I'll probably download Python and check in a bit), but in the
expected case (where JES is the only programming editor on the system),
it works just as intended. (It does coexist peacefully with Notepad++.)

Support for this was enabled by a new JES.exe launcher, contributed by
Stephen Bryant. The old one doesn't support being launched with any
working directory other than the JES root, and it doesn't pass its
command line arguments on to JES.bat. Stephen's new launcher does both.
It's also smaller and faster, and can be built with a more minimal
toolchain.