make jes.sh XDG base directory spec compliant
On not-Mac platforms, jes.sh uses the XDG standard ~/.config directory
for user config files, and ~/.cache for cache files. But, the spec says
that if there's an XDG_CONFIG_HOME or XDG_CACHE_HOME, you have to use
that. I didn't implement it earlier because I was lazy and didn't feel
like making my 'if' statements longer, but now that I had all the XDG
specs open anyway, I figured it would be a good time.