automatically clamp font size to 8..72 on load
West Point reported a bug where occasionally the font size would be
set to 0 in the user's config file, and JES would accept the value and
make all their text invisible. (Since 0 isn't an option in the font
size dropdown box, I'm assuming it's an artifact of the old config
file syntax.)

Nevertheless, JESConfig's afterLoad method now adjusts the font size
on load. If it's 0 (indicating a missing key or [now] syntax error), it
resets it to 12 (the default). Otherwise, it just pulls it to the
nearest extreme of the range 8..72.

While adjusting this, I noticed that on sub-12 font sizes, the Command
Window's insertion point is larger than the line height, adding an
artificial margin to the input line. While this could be considered a
bug, this behavior does have the advantage of visually distinguishing
the input line from the rest (useful at these small font sizes).
So for the time being, I'm going to categorize it as an accidental
feature.

I also noticed that syntax errors in integer configuration keys cause
exceptions. I'm fairly sure that the old behavior was to return 0, and
if not, I still don't think any of the getIntegerProperty calls have
NumberFormatException handlers around them. So, it was adjusted to
catch NumberFormatException and return 0.