bundle a copy of jMusic
Now, a jMusic jar (containing the "jm" package) is included in the JES
dependencies, and the build process compiles all the instruments from
the jMusic instruments download archive and includes them as well.
There's no special integration or media.py functions, but you can
import jm and the stuff inside it, and use that to make music.

There's not currently any documentation or help files, aside from a
hotCrossBuns.py demo included in the demos folder. (It's the only song
I ever learned to play on piano, so it's the only song I have the score
to memorized.)