add "Stop" buttons to requestX/showX dialogs
Another comment from West Point was that commonly students would create
infinite loops of dialog boxes, which could not be stopped, because the
"Stop" button was blocked by the modal-ness. This mitigates that by
adding a "Stop" button to each dialog box created from user code.
Clicking it stops the program.