don't type function names from "JES Functions" menu
Previously, JES typed the name of a function followed by a left paren
whenever it was selected from the "JES Functions" menu. This was
reported as actually being more confusing for students, because when
consulting the JES Functions menu, they usually haven't decided whether
to use it or not yet. (Also it would frequently type it in
syntactically inappropriate places.)