rewrite the new/open/save code
Now, it lives in an isolated package (jes.gui.filemanager). It uses the
methodAction pattern like commandWindow and friends, and the methods
are clearly split between the "model" and "view" layers.

This rewrite was intended to accomodate some future changes requested
by West Point, and to clean up the error handling. It also allowed a
bunch of code reuse and simplification of the control flow.