audit and add third-party licensing inforamtion
This lives in a JESDependencies.txt file, which includes a list of
all the third-party components in JES with the attendant licenses.

Fortunately, all our code is legally redistributable (though I have to
admit that jmf.jar made me a nervous), but some of the licensing and
versioning information isn't 100% clear, and I had to make
some inferences.