rewrite printing code
A couple of changes I made earlier slightly broke printing. This commit
gets it back into shape, and makes several improvements (like removing
the printing GUI code from JESUI, and keeping it with the filemanager).

The actual print code itself ended up being rewritten, because it was
super messy -- it took me about a half hour to understand the purposes
of all the variables and whatnot. Now the code is much better commented
and generates a nicer layout to boot. It also supports out-of-order
printing better.