created a new prompt for when saving is required
The continueAfterSaving prompt was written with the assumption that
not saving is OK. However, there are other cases where the user HAS
to save the file, or bad stuff will happen when they try to load or
print it.

So, the old prompt was renamed to continueAfterSavingOrDiscarding,
still having Yes/No/Cancel options, and the new continueAfterSaving
only presents options for Yes (save and continue) and No (don't save
and stop).