fix movie frame sequencing and ordering
Most notably, the Movie class was configured to sort the frames before
trying to play them, instead of just accepting whatever order the OS
gives. I noticed the symptoms of this bug earlier, but thought it was
a bug in my movie program.

But, the frame sequencer tool also had some minor improvements.
It was moved into the jes.tools package instead of a standalone .py,
and the Movie class was changed to a new-style class so Explorers would
detect it properly.