ultimate victory over JESConstants.py
All the remaining constants were only used in one place, and most of
those files already had a constant bank near the top anyway.
So, all the constants were excised from JESConstants, and the ones that
were still in use were put in the files they are used in.
Then, JESConstants.py was destroyed.