remove LogBuffer and associated UI
There were already faults in the implementation, some of which were
created when I removed logging hooks as I rewrote code, but some of
which were already present in JES 4.3. The information collected
seemed a bit haphazard, and since it suggested that the information
was mostly for debugging, I decided it was safe to remove.