make requestString Unicode again
I thought the str() call in requestString was a bug because it would
turn Nones into 'None's. It was actually designed to convert from
unicode strings to str strings, and the creation of 'None' was an
unintended side effect. As a result, requestString was returning
unicodes. I put the str back, inside an "is not None" check.

(Also I think I put that str there to begin with.)