add a basic plugin system
The JES plugin system, as it stands, is basically just a fancy way to
install extra JAR's full of Java and Python libraries. It provides
menu options for "Install Plugin" and "Manage Plugins," which let you
copy in plugins.

There's also a Plugin Creation Kit included in the source code
repository, which you can copy to make a new plugin. Just edit a couple
of files, add your stuff, run 'ant', and it creates a plugin for you.