allow JES's name to be customized
In many places, you can now override the name "JES" from the Ant file.
(It does this by reading the name from JESVersion.) This is not
everywhere, many parts of JES still have "JES" hardcoded. This is just
enough to allow users to create custom builds of JES, and give them
identifiable names.

(The names do have to have "JES" in them, though.)