refine file manager's UI
First, I added the JFileChooser on the file manager to the list of
things to refresh when changing the skin. This ensures that it matches
the current L&F.

Second, I had it reset the current directory and selected file when
opening the file chooser whenever a file is open. This leads to more
predictable behavior.

And third, I added a brief rant about how terrible the GTK+
JFileChooser is, and officially declared it to be Oracle's problem.